modify column data type

+++ init.sql
@@ -8,7 +8,7 @@ USE naezuo;
 
 -- 테이블 생성
 CREATE TABLE `user` (
-                        `id` int NOT NULL AUTO_INCREMENT,
+                        `id` bigint NOT NULL AUTO_INCREMENT,
                         `user_id` varchar(50) NOT NULL COMMENT '로그인 아이디',
                         `pw` varchar(255) NOT NULL,
                         `age` int NOT NULL,
@@ -20,7 +20,7 @@ CREATE TABLE `user` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
 
 CREATE TABLE `goods` (
-                         `id` int NOT NULL AUTO_INCREMENT,
+                         `id` bigint NOT NULL AUTO_INCREMENT,
                          `name` varchar(100) NOT NULL,
                          `cost` decimal(10,2) NOT NULL,
                          `total_amount` decimal(15,2) NOT NULL COMMENT '총 구매금액',
@@ -29,9 +29,9 @@ CREATE TABLE `goods` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
 
 CREATE TABLE `funding` (
-                           `id` int NOT NULL AUTO_INCREMENT,
-                           `user_id` int NOT NULL,
-                           `goods_id` int NOT NULL,
+                           `id` bigint NOT NULL AUTO_INCREMENT,
+                           `user_id` bigint NOT NULL,
+                           `goods_id` bigint NOT NULL,
                            `amount` decimal(15,2) NOT NULL COMMENT '구매금액',
                            `count` int NOT NULL COMMENT '구매개수',
                            PRIMARY KEY (`id`),
